Move plurals cldr_serde
